+++ packages/docusaurus-theme/src/components/index.ts
@@ -9,5 +9,4 @@
 export { createDemo, Demo, type DemoProps } from './demo';
 export { Guideline, GuidelineText } from './guideline';
 export { PropTable } from './prop_table';
-export { StorybookLink } from './storybook_link';
 export { AppThemeContext } from './theme_context';
