--- packages/docusaurus-theme/src/components/storybook_link/has_pr_root_url.ts
@@ -1,13 +1,0 @@
-export const hasPrRootUrl = (
-  obj: unknown
-): obj is { customFields: { prRootUrl: string } } => {
-  return (
-    typeof obj === 'object' &&
-    obj !== null &&
-    'customFields' in obj &&
-    typeof obj.customFields === 'object' &&
-    obj.customFields !== null &&
-    'prRootUrl' in obj.customFields &&
-    typeof obj.customFields.prRootUrl === 'string'
-  );
-};
